Move all main controller route methods to a concern
