+++ app/controllers/candidate_interface/new_references/accept_offer/email_address_controller.rb
@@ -2,22 +2,7 @@ module CandidateInterface
   module NewReferences
     class AcceptOffer::EmailAddressController < EmailAddressController
       include AcceptOfferConfirmReferences
-
-      def references_email_address_path
-        candidate_interface_accept_offer_new_references_email_address_path(
-          application_choice,
-          @reference.id,
-        )
-      end
-
-      def edit_email_address_path
-        candidate_interface_accept_offer_new_references_edit_email_address_path(
-          application_choice,
-          @reference.id,
-          return_to: params[:return_to],
-        )
-      end
-      helper_method :edit_email_address_path
+      include AcceptOfferNewReferencesPath
 
       def previous_path
         candidate_interface_accept_offer_new_references_name_path(
