+++ app/controllers/candidate_interface/new_references/accept_offer/relationship_controller.rb
@@ -2,23 +2,7 @@ module CandidateInterface
   module NewReferences
     class AcceptOffer::RelationshipController < RelationshipController
       include AcceptOfferConfirmReferences
-
-      def references_relationship_path
-        candidate_interface_accept_offer_new_references_relationship_path(
-          application_choice,
-          @reference.id,
-        )
-      end
-      helper_method :references_relationship_path
-
-      def edit_relationship_path
-        candidate_interface_accept_offer_new_references_edit_relationship_path(
-          application_choice,
-          @reference.id,
-          return_to: params[:return_to],
-        )
-      end
-      helper_method :edit_relationship_path
+      include AcceptOfferNewReferencesPath
 
       def previous_path
         candidate_interface_accept_offer_new_references_email_address_path(
