+++ app/controllers/candidate_interface/new_references/accept_offer/type_controller.rb
@@ -2,33 +2,7 @@ module CandidateInterface
   module NewReferences
     class AcceptOffer::TypeController < TypeController
       include AcceptOfferConfirmReferences
-
-      def references_type_path
-        candidate_interface_accept_offer_new_references_type_path(
-          application_choice,
-          params[:referee_type],
-          params[:id],
-        )
-      end
-      helper_method :references_type_path
-
-      def reference_new_type_path
-        candidate_interface_accept_offer_new_references_type_path(
-          application_choice,
-          params[:referee_type],
-          params[:id],
-        )
-      end
-      helper_method :reference_new_type_path
-
-      def reference_edit_type_path
-        candidate_interface_accept_offer_new_references_edit_type_path(
-          application_choice,
-          @reference.id,
-          return_to: params[:return_to],
-        )
-      end
-      helper_method :reference_edit_name_path
+      include AcceptOfferNewReferencesPath
 
       def previous_path
         candidate_interface_accept_offer_path(application_choice)
